+++ resources/ts/index.tsx
@@ -1,7 +1,7 @@
+import './bootstrap';
 import { createInertiaApp } from '@inertiajs/react';
 import React from 'react';
 import { createRoot } from 'react-dom/client';
-import Layout from './Pages/Layout';
 import "bootstrap";
 import '../scss/__toast.scss';
 import 'react-toastify/dist/ReactToastify.css';