+++ resources/views/app.blade.php
@@ -6,12 +6,12 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <meta name="csrf-token" content="{{ csrf_token() }}">
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.5.2/css/all.min.css" integrity="sha512-SnH5WK+bZxgPHs44uWIX+LLJAJ9/2PkPKZ5QiAj6Ta86w+fsb2TkcmfRyVX3pBnMFcV7oQPJkl9QevSCWr3W6A==" crossorigin="anonymous" referrerpolicy="no-referrer" />
-    <link rel="shortcut icon" href="favicon.png" type="image/png">
+    <link rel="shortcut icon" href="/favicon.png" type="image/png">
 
 
     @inertiaHead
     @viteReactRefresh
-    @vite('resources/ts/bootstrap.ts')
+
     @vite('resources/scss/style.scss')
     @vite('resources/ts/index.tsx')
     @routes
@@ -19,8 +19,6 @@
 
 <body>
     @include('Fragments.loader')
-
-
     @inertia
 </body>